borrowedBooksByUserView

+++ src/main/java/pl/marta/kopp/library/ReturnController.java
@@ -6,6 +6,9 @@ import pl.marta.kopp.domain.borrow.Borrow;
 import pl.marta.kopp.persistence.BookStorage;
 import pl.marta.kopp.persistence.BorrowStorage;
 
+import java.util.ArrayList;
+import java.util.List;
+
 public class ReturnController {
     private final BookStorage bookStorage;
     private final BorrowStorage borrowStorage;
@@ -19,17 +22,34 @@ public class ReturnController {
         if (borrowStorage.isExistsUserId(userId)) {
             if (borrowStorage.isExistsBookId(bookId)) {
                 deleteBorrow(bookId);
-                bookStorage.setBorrow(bookId,false);
+                bookStorage.setBorrow(bookId, false);
                 return Response.aSuccessfulResponse();
             } else return Response.aFailureResponse("You don't have any borrowed book.");
         } else return Response.aFailureResponse("Invalid Book Id");
     }
 
     private void deleteBorrow(long bookId) {
-        Borrow borrow = borrowStorage.getByBookId(bookId);
-        borrowStorage.delete(borrow.getId());
+        if (borrowStorage.isExistsBookId(bookId)) {
+            Borrow borrow = borrowStorage.getByBookId(bookId);
+            borrowStorage.delete(borrow.getId());
+        } else throw new BorrowDoesNotExistsException();
     }
 
+    private List<Borrow> getBorrowByUserId(long userId) {
+        if (borrowStorage.isExistsUserId(userId)) {
+            return borrowStorage.getByUserId(userId);
+        } else throw new BorrowDoesNotExistsException();
+    }
+
+    public List<Book> getBorrowedBooksByUserId(long userId) {
+        List<Borrow> borrows = getBorrowByUserId(userId);
+        List<Book> books = new ArrayList<>();
+        for (Borrow b : borrows) {
+            Book bookById = bookStorage.getById(b.getBookId());
+            books.add(bookById);
+        }
+        return books;
+    }
 }
